Prototyyppi Serial monitorin avulla

+++ SpedenSpelit/buttons.h
@@ -8,7 +8,7 @@ int convert (int);
 const byte firstPin = 2; // First PinChangeInterrupt on D-bus
 const byte lastPin =  5; // Last PinChangeInterrupt on D-bus
 
-extern volatile int buttonPressed;
+extern volatile byte buttonPressed;
 extern bool canPress;
 extern bool gameRunning;
